orthoprint/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
trimesh[all]==4.3.2
numpy==1.26.4
scipy==1.13.0
shapely==2.0.4
python-multipart==0.0.9
pydantic==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
